requirements.txt
Django>=3.2,<4.0
transformers
torch
pillow
requests
gunicorn
